+++ src/components/category/category.js
@@ -4,7 +4,7 @@ import './category.scss';
 
 const { results } = categories;
 
-const category = () => {
+export const CategoryCard = () => {
   return (
     <section className="section-wrapper">
       <h1 className="h1-title mb-15">Categories</h1>
@@ -22,7 +22,39 @@ const category = () => {
         })}
       </div>
     </section>
-  )
+  );
 }
 
-export default category
+
+
+export const CategoryList = ({setCategoriesState, categories}) => {
+
+  const handleCheck = (e, categoryName) => {
+    const isSelected = e.target.checked;
+    setCategoriesState(
+      {
+        ...categories,
+        [categoryName]: isSelected,
+      }
+    );
+  };
+
+  return (
+    <>
+      <span>Categories</span>
+        { results.map(({id, data:{name}}) => {
+          return (
+            <div className="form-checkbox" key={id}>
+              <input
+                type="checkbox"
+                id={id}
+                value={name}
+                onChange={(e) => {handleCheck(e, e.target.value)}}
+              />
+              <label htmlFor={id}>{name}</label>
+            </div>
+          );
+        })}
+    </>
+  )
+}
